@@ -71,54 +76,59 @@
     "passport": "0.3.2",
     "passport-facebook": "^2.1.1",
     "passport-local": "^1.0.0",
-    "pug": "^2.0.0-beta6",
-    "request": "^2.78.0",
+    "pug": "^2.0.0-rc.4",
+    "request": "^2.83.0",
+    "request-promise": "^4.2.2",
     "sharp": "^0.18.4",
-    "twilio": "^3.8.1",
+    "stream-buffers": "^3.0.1",
+    "twilio": "^3.11.0",
     "typed-promisify": "^0.4.0",
     "underscore": "^1.8.3",
     "web-request": "^1.0.7"
   },
   "devDependencies": {
-    "@types/async": "^2.0.40",
+    "@types/async": "^2.0.45",
     "@types/bcrypt-nodejs": "0.0.30",
-    "@types/body-parser": "^1.16.2",
+    "@types/bluebird": "^3.5.18",
+    "@types/body-parser": "^1.16.8",
     "@types/compression": "0.0.33",
     "@types/connect-mongo": "0.0.32",
-    "@types/dotenv": "^2.0.20",
+    "@types/dotenv": "^2.0.23",
     "@types/errorhandler": "0.0.30",
-    "@types/express": "^4.0.35",
+    "@types/express": "^4.0.39",
     "@types/express-session": "0.0.32",
     "@types/facebook-js-sdk": "^2.8.3",
     "@types/file-type": "^5.2.1",
     "@types/helmet": "0.0.37",
     "@types/jest": "^19.2.2",
     "@types/jimp": "^0.2.1",
-    "@types/jquery": "^2.0.41",
-    "@types/lodash": "^4.14.63",
+    "@types/jquery": "^2.0.48",
+    "@types/lodash": "^4.14.91",
     "@types/method-override": "0.0.31",
-    "@types/mongodb": "^2.1.43",
-    "@types/morgan": "^1.7.32",
-    "@types/multer": "^1.3.5",
-    "@types/node": "^7.0.12",
+    "@types/mongodb": "^2.2.17",
+    "@types/morgan": "^1.7.35",
+    "@types/multer": "^1.3.6",
+    "@types/node": "^7.0.50",
     "@types/nodemailer": "^1.3.33",
-    "@types/passport": "^0.3.3",
-    "@types/passport-facebook": "^2.1.3",
+    "@types/passport": "^0.3.5",
+    "@types/passport-facebook": "^2.1.7",
     "@types/pug": "^2.0.4",
-    "@types/request": "0.0.45",
-    "@types/sharp": "^0.17.4",
-    "@types/supertest": "^2.0.0",
+    "@types/request": "^2.0.6",
+    "@types/request-promise": "^4.1.39",
+    "@types/sharp": "^0.17.6",
+    "@types/stream-buffers": "^3.0.2",
+    "@types/supertest": "^2.0.4",
     "@types/twilio": "0.0.9",
-    "@types/underscore": "^1.8.3",
-    "concurrently": "^3.4.0",
+    "@types/underscore": "^1.8.6",
+    "concurrently": "^3.5.1",
     "jest": "^19.0.2",
     "json-loader": "^0.5.7",
-    "node-sass": "^4.5.2",
-    "nodemon": "^1.11.0",
-    "shelljs": "^0.7.7",
+    "node-sass": "^4.7.2",
+    "nodemon": "^1.13.3",
+    "shelljs": "^0.7.8",
     "supertest": "^2.0.1",
     "ts-jest": "^19.0.8",
-    "tslint": "^5.0.0",
-    "typescript": "^2.4.0"
+    "tslint": "^5.8.0",
+    "typescript": "^2.6.2"
   }
 }
