@@ -1,34 +1,17 @@
-import * as admin from "firebase-admin";
-import {NotificationLog} from "../models/notification-logs";
-import * as fs from "fs";
+import { NotificationLog } from "../models/notification-logs";
+import { firebaseClient } from "./firebase-client";
+import { User } from "../models/user";
 
 export class NotificationSender {
-    private static _initialized = false;
     protected _tokens: string[];
     protected _title: string = "";
     protected _type: number = 0;
     protected _message: string = "";
     protected _payload: any = {};
+    protected _sentFromCms = false;
 
     constructor(tokens: string|string[]) {
-        if ( ! NotificationSender._initialized ) {
-            const firebasePrivateCertificate = fs.readFileSync(__dirname + "/../../firebase.pem", "utf8");
-
-            const options = {
-                credential: admin.credential.cert({
-                    projectId: process.env.FIREBASE_PROJECT_ID,
-                    privateKey: firebasePrivateCertificate,
-                    clientEmail: process.env.FIREBASE_CLIENT_EMAIL
-                }),
-                databaseURL: process.env.FIREBASE_DATABASE_URL
-            };
-
-            admin.initializeApp(options);
-
-            NotificationSender._initialized = true;
-        }
-
-        this._tokens = tokens instanceof Array ? tokens : [tokens];
+        this._tokens = tokens instanceof Array ? tokens : [...new Set(tokens)];
     }
 
     title(title: string) {
@@ -80,15 +63,49 @@ export class NotificationSender {
             const notificationLog = new NotificationLog;
             notificationLog.tokens = this._tokens;
             notificationLog.payload = payload;
+            notificationLog.sentByAdministrator = this._sentFromCms;
 
-            admin
+            firebaseClient()
                 .messaging()
                 .sendToDevice(that._tokens, payload)
                 .then(async (response) => {
                     notificationLog.success = response;
                     resolve(response);
 
-                    await notificationLog.save();
+                    const promisesList: Array<any> = [
+                        notificationLog.save()
+                    ];
+
+                    if ( response.results ) {
+                        const removeTokens: string[] = [];
+                        const resultsWithIndexes = response.results.map((result, index) => ({ result, index }));
+
+                        for ( const {result, index} of resultsWithIndexes ) {
+                            if ( result.error && result.error.code === "messaging/registration-token-not-registered" ) {
+                                removeTokens.push(this._tokens[index]);
+                            }
+                        }
+
+                        if ( removeTokens.length > 0 ) {
+                            const conditions = {
+                                "tokens.firebaseToken": { $in: removeTokens }
+                            };
+
+                            const update = {
+                                "$set": {
+                                    "tokens.$.firebaseToken": ""
+                                }
+                            };
+
+                            const options = {
+                                multi: true
+                            };
+
+                            promisesList.push(User.update(conditions, update, options));
+                        }
+                    }
+
+                    await Promise.all(promisesList);
                 })
                 .catch(async (error) => {
                     notificationLog.error = error;
